requirements.txt
torch
numpy
pandas
tqdm
kagglehub
sacrebleu
evaluate
rouge_score
tokenizers
pyyaml
tensorboard
